# Conflicts:
#	CHANGELOG.md
#	test/baselines/TS_Prettier/package.json
#	test/baselines/adapter_TS_React/package.json

+++ src/lib/core/questions.ts
@@ -780,7 +780,7 @@ export const questionGroups: QuestionGroup[] = [
 				name: "quotes",
 				label: "Quotes",
 				message: "Do you prefer double or single quotes?",
-				initial: "double",
+				initial: "single",
 				choices: ["double", "single"],
 				migrate: async ctx => ((await ctx.analyzeCode('"', "'")) ? "double" : "single"),
 			},
@@ -1197,15 +1197,8 @@ export async function formatAnswers(answers: Record<string, any>): Promise<Recor
 		}
 
 		// Apply default value from initial if answer is missing and not optional
-		if (
-			answers[q.name as string] == undefined &&
-			!q.optional &&
-			q.initial !== undefined
-		) {
-			answers[q.name as string] =
-				typeof q.initial === "function"
-					? q.initial(answers)
-					: q.initial;
+		if (answers[q.name as string] == undefined && !q.optional && q.initial !== undefined) {
+			answers[q.name as string] = typeof q.initial === "function" ? q.initial(answers) : q.initial;
 		}
 
 		// Apply an optional transformation
